Reorganized the project
