+++ source/model/player.py
@@ -1,4 +1,3 @@
-import config
 from enum import IntEnum
